% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_atac.R
\name{ATACAdapter}
\alias{ATACAdapter}
\title{ATAC Adapter}
\description{
Adapter for storing and retrieving ATAC-layer objects.
}
\details{
Supports explicit ATAC marker classes/metadata and ChromatinAssay objects.
}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{ATACAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ATACAdapter-name}{\code{ATACAdapter$name()}}
\item \href{#method-ATACAdapter-priority}{\code{ATACAdapter$priority()}}
\item \href{#method-ATACAdapter-put}{\code{ATACAdapter$put()}}
\item \href{#method-ATACAdapter-get}{\code{ATACAdapter$get()}}
\item \href{#method-ATACAdapter-components}{\code{ATACAdapter$components()}}
\item \href{#method-ATACAdapter-exists}{\code{ATACAdapter$exists()}}
\item \href{#method-ATACAdapter-list_names}{\code{ATACAdapter$list_names()}}
\item \href{#method-ATACAdapter-can_handle}{\code{ATACAdapter$can_handle()}}
\item \href{#method-ATACAdapter-clone}{\code{ATACAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ATACAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-ATACAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ATACAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ATACAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-ATACAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ATACAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ATACAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-ATACAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ATACAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ATACAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-ATACAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ATACAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ATACAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-ATACAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ATACAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ATACAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-ATACAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ATACAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ATACAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-ATACAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ATACAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ATACAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-ATACAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ATACAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ATACAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ATACAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ATACAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
