% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_chip.R
\name{ChIPAdapter}
\alias{ChIPAdapter}
\title{ChIP Adapter}
\value{
An R6 generator for a \code{LakeAdapter} subclass that
  serializes and restores objects of this omics layer.
}
\description{
Adapter for storing and retrieving ChIP-layer objects.
}
\details{
Supports explicit ChIP marker classes/metadata.
}
\examples{
adapter <- ChIPAdapter$new()
class(adapter)
}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{ChIPAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChIPAdapter-name}{\code{ChIPAdapter$name()}}
\item \href{#method-ChIPAdapter-priority}{\code{ChIPAdapter$priority()}}
\item \href{#method-ChIPAdapter-put}{\code{ChIPAdapter$put()}}
\item \href{#method-ChIPAdapter-get}{\code{ChIPAdapter$get()}}
\item \href{#method-ChIPAdapter-components}{\code{ChIPAdapter$components()}}
\item \href{#method-ChIPAdapter-exists}{\code{ChIPAdapter$exists()}}
\item \href{#method-ChIPAdapter-list_names}{\code{ChIPAdapter$list_names()}}
\item \href{#method-ChIPAdapter-can_handle}{\code{ChIPAdapter$can_handle()}}
\item \href{#method-ChIPAdapter-clone}{\code{ChIPAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChIPAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-ChIPAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChIPAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChIPAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-ChIPAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChIPAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChIPAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-ChIPAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChIPAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChIPAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-ChIPAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChIPAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChIPAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-ChIPAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChIPAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChIPAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-ChIPAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChIPAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChIPAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-ChIPAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChIPAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChIPAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-ChIPAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChIPAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChIPAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChIPAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChIPAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
