% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_chromatograms.R
\name{ChromatogramsAdapter}
\alias{ChromatogramsAdapter}
\title{Chromatograms Adapter}
\description{
Adapter for storing and retrieving
\code{Chromatograms} objects.
}
\details{
This adapter currently guarantees full-fidelity roundtrip by storing the
complete object and manifest as internal components.
}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{ChromatogramsAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChromatogramsAdapter-name}{\code{ChromatogramsAdapter$name()}}
\item \href{#method-ChromatogramsAdapter-can_handle}{\code{ChromatogramsAdapter$can_handle()}}
\item \href{#method-ChromatogramsAdapter-priority}{\code{ChromatogramsAdapter$priority()}}
\item \href{#method-ChromatogramsAdapter-put}{\code{ChromatogramsAdapter$put()}}
\item \href{#method-ChromatogramsAdapter-get}{\code{ChromatogramsAdapter$get()}}
\item \href{#method-ChromatogramsAdapter-components}{\code{ChromatogramsAdapter$components()}}
\item \href{#method-ChromatogramsAdapter-exists}{\code{ChromatogramsAdapter$exists()}}
\item \href{#method-ChromatogramsAdapter-list_names}{\code{ChromatogramsAdapter$list_names()}}
\item \href{#method-ChromatogramsAdapter-clone}{\code{ChromatogramsAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChromatogramsAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-ChromatogramsAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromatogramsAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChromatogramsAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-ChromatogramsAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromatogramsAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChromatogramsAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-ChromatogramsAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromatogramsAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChromatogramsAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-ChromatogramsAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromatogramsAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChromatogramsAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-ChromatogramsAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromatogramsAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChromatogramsAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-ChromatogramsAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromatogramsAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChromatogramsAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-ChromatogramsAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromatogramsAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChromatogramsAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-ChromatogramsAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromatogramsAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChromatogramsAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChromatogramsAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromatogramsAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
