% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{coalesce}
\alias{coalesce}
\title{Coalesce - return first non-NA value}
\usage{
coalesce(...)
}
\arguments{
\item{...}{Vectors to coalesce}
}
\value{
Vector with first non-NA values
}
\description{
Thin wrapper around \code{dplyr::coalesce()} re-exported for convenience.
Note: this will mask \code{dplyr::coalesce} when both packages are loaded.
}
\examples{
x <- c(NA, 2, NA)
y <- c(1, NA, 3)
coalesce(x, y) # Returns c(1, 2, 3)
}
