% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{contains_str}
\alias{contains_str}
\title{Check if string contains a substring}
\usage{
contains_str(x, substring)
}
\arguments{
\item{x}{Character vector}

\item{substring}{Substring to search for}
}
\value{
Logical vector
}
\description{
Check if string contains a substring
}
\examples{
text <- c("hello world", "goodbye", "hello there")
text[contains_str(text, "hello")]
}
