% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{create_pipeline}
\alias{create_pipeline}
\title{Create a pipeline of tracked operations}
\usage{
create_pipeline(lake, name)
}
\arguments{
\item{lake}{Lake instance}

\item{name}{Name for this pipeline}
}
\value{
A Pipeline object
}
\description{
Defines a sequence of operations that will be tracked as a unit.
}
\examples{
if (FALSE) {
    lake <- Lake$new("project")

    pipeline <- create_pipeline(lake, "preprocessing")
    pipeline$
        step("load", function() read.csv("data.csv"))$
        step("clean", function(data) na.omit(data))$
        step("normalize", function(data) scale(data))$
        run()

    # All steps are recorded in lineage
    lake$tree("preprocessing.normalize")
}
}
