% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{lake}
\alias{lake}
\title{Get the current default lake}
\usage{
lake()
}
\value{
The default Lake object
}
\description{
Get the current default lake
}
\examples{
use_lake("ex_lake", root = tempfile())
l <- lake()
l$put("data", data.frame(x = 1:3))
}
