% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lake.R
\name{DependencyTracker}
\alias{DependencyTracker}
\title{Dependency Tracker}
\description{
Tracks data dependencies during operations with version-aware lineage
}
\keyword{Record}
\keyword{a}
\keyword{internal}
\keyword{operation}
\keyword{read}
\keyword{reference}
\keyword{version}
\keyword{with}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DependencyTracker-new}{\code{DependencyTracker$new()}}
\item \href{#method-DependencyTracker-track_read}{\code{DependencyTracker$track_read()}}
\item \href{#method-DependencyTracker-start_write}{\code{DependencyTracker$start_write()}}
\item \href{#method-DependencyTracker-end_write}{\code{DependencyTracker$end_write()}}
\item \href{#method-DependencyTracker-current_reads}{\code{DependencyTracker$current_reads()}}
\item \href{#method-DependencyTracker-current_read_names}{\code{DependencyTracker$current_read_names()}}
\item \href{#method-DependencyTracker-clear}{\code{DependencyTracker$clear()}}
\item \href{#method-DependencyTracker-clone}{\code{DependencyTracker$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DependencyTracker-new"></a>}}
\if{latex}{\out{\hypertarget{method-DependencyTracker-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyTracker$new(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DependencyTracker-track_read"></a>}}
\if{latex}{\out{\hypertarget{method-DependencyTracker-track_read}{}}}
\subsection{Method \code{track_read()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyTracker$track_read(name, ref = "@latest")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the data being read}

\item{\code{ref}}{Version reference (e.g., "@latest", "@tag(v1)")
Start tracking a write operation
End tracking and return dependencies with version info}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DependencyTracker-start_write"></a>}}
\if{latex}{\out{\hypertarget{method-DependencyTracker-start_write}{}}}
\subsection{Method \code{start_write()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyTracker$start_write(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DependencyTracker-end_write"></a>}}
\if{latex}{\out{\hypertarget{method-DependencyTracker-end_write}{}}}
\subsection{Method \code{end_write()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyTracker$end_write()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of dependency objects with name and ref
Get current tracked reads with version info
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DependencyTracker-current_reads"></a>}}
\if{latex}{\out{\hypertarget{method-DependencyTracker-current_reads}{}}}
\subsection{Method \code{current_reads()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyTracker$current_reads()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of dependency objects with name and ref
Get current tracked reads as character vector (for backward compatibility)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DependencyTracker-current_read_names"></a>}}
\if{latex}{\out{\hypertarget{method-DependencyTracker-current_read_names}{}}}
\subsection{Method \code{current_read_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyTracker$current_read_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character vector of dependency names
Clear all tracking
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DependencyTracker-clear"></a>}}
\if{latex}{\out{\hypertarget{method-DependencyTracker-clear}{}}}
\subsection{Method \code{clear()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyTracker$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DependencyTracker-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DependencyTracker-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyTracker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
