% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{deps}
\alias{deps}
\title{Get dependencies from the default lake}
\usage{
deps(name, direction = "up")
}
\arguments{
\item{name}{Data name}

\item{direction}{"up" or "down"}
}
\value{
Dependencies data frame
}
\description{
Get dependencies from the default lake
}
\examples{
use_lake("ex_deps", root = tempfile())
put("raw", data.frame(x = 1:3))
deps("raw")
}
