% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_bench.R
\name{.ol_eval_cleanup_project}
\alias{.ol_eval_cleanup_project}
\title{Cleanup evaluation project}
\usage{
.ol_eval_cleanup_project(root, project_name)
}
\description{
Cleanup evaluation project
}
\keyword{internal}
