% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_bench.R
\name{.ol_eval_get_project_size}
\alias{.ol_eval_get_project_size}
\title{Get project size in bytes}
\usage{
.ol_eval_get_project_size(root, lake)
}
\description{
Get project size in bytes
}
\keyword{internal}
