% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_bench.R
\name{.ol_eval_query_spec_filter_group}
\alias{.ol_eval_query_spec_filter_group}
\title{Query spec for W1-2: filter + group + summarize}
\usage{
.ol_eval_query_spec_filter_group(tbl)
}
\arguments{
\item{tbl}{A lazy table reference (tbl_lazy or lake_tbl)}
}
\value{
Lazy query result
}
\description{
Query spec for W1-2: filter + group + summarize
}
\keyword{internal}
