% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_bench.R
\name{.ol_eval_query_spec_filter_select}
\alias{.ol_eval_query_spec_filter_select}
\title{Query spec for W1-1: simple filter + select}
\usage{
.ol_eval_query_spec_filter_select(
  tbl,
  filter_col = "num1",
  threshold = 50,
  select_cols = c("id", "num1", "cat1")
)
}
\arguments{
\item{tbl}{A lazy table reference}

\item{filter_col}{Column to filter on}

\item{threshold}{Filter threshold}

\item{select_cols}{Columns to select}
}
\value{
Lazy query result
}
\description{
Query spec for W1-1: simple filter + select
}
\keyword{internal}
