% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_plot.R
\name{.ol_eval_records_to_df}
\alias{.ol_eval_records_to_df}
\title{Convert JSONL records to data frame}
\usage{
.ol_eval_records_to_df(records)
}
\description{
Convert JSONL records to data frame
}
\keyword{internal}
