% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_bench.R
\name{.ol_eval_W0_1_put_table}
\alias{.ol_eval_W0_1_put_table}
\title{W0-1: put(table) benchmark}
\usage{
.ol_eval_W0_1_put_table(
  lake,
  size_cfg,
  size_name,
  config,
  env_info,
  output_file
)
}
\description{
W0-1: put(table) benchmark
}
\keyword{internal}
