% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_bench.R
\name{.ol_eval_W0_2_put_object}
\alias{.ol_eval_W0_2_put_object}
\title{W0-2: put(object) benchmark}
\usage{
.ol_eval_W0_2_put_object(lake, size_name, config, env_info, output_file)
}
\description{
W0-2: put(object) benchmark
}
\keyword{internal}
