% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_bench.R
\name{.ol_eval_W0_3_tag_snap}
\alias{.ol_eval_W0_3_tag_snap}
\title{W0-3: tag/snap benchmark with storage breakdown (P0-3: C2)}
\usage{
.ol_eval_W0_3_tag_snap(lake, size_name, config, env_info, output_file)
}
\description{
W0-3: tag/snap benchmark with storage breakdown (P0-3: C2)
}
\keyword{internal}
