% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_bench.R
\name{.ol_eval_W1_1_get_pushdown}
\alias{.ol_eval_W1_1_get_pushdown}
\title{W1-1: get with where/select pushdown}
\usage{
.ol_eval_W1_1_get_pushdown(lake, size_name, config, env_info, output_file)
}
\description{
W1-1: get with where/select pushdown
}
\keyword{internal}
