% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_bench.R
\name{.ol_eval_W1_2_ref_dplyr}
\alias{.ol_eval_W1_2_ref_dplyr}
\title{W1-2: ref + dplyr pipeline (uses shared query spec)}
\usage{
.ol_eval_W1_2_ref_dplyr(lake, size_name, config, env_info, output_file)
}
\description{
W1-2: ref + dplyr pipeline (uses shared query spec)
}
\keyword{internal}
