% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_bench.R
\name{.ol_eval_W1_3_joins}
\alias{.ol_eval_W1_3_joins}
\title{W1-3: Join benchmarks}
\usage{
.ol_eval_W1_3_joins(lake, size_name, config, env_info, output_file)
}
\description{
W1-3: Join benchmarks
}
\keyword{internal}
