% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_bench.R
\name{.ol_eval_W2_1_deps_tree}
\alias{.ol_eval_W2_1_deps_tree}
\title{W2-1: deps/tree benchmark}
\usage{
.ol_eval_W2_1_deps_tree(lake, config, env_info, output_file)
}
\description{
W2-1: deps/tree benchmark
}
\keyword{internal}
