% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_bench.R
\name{.ol_eval_W2_2_impact}
\alias{.ol_eval_W2_2_impact}
\title{W2-2: impact analysis benchmark}
\usage{
.ol_eval_W2_2_impact(lake, config, env_info, output_file)
}
\description{
W2-2: impact analysis benchmark
}
\keyword{internal}
