% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{.ol_init_backend}
\alias{.ol_init_backend}
\title{Initialize DuckDB backend (internal)}
\usage{
.ol_init_backend(
  project,
  engine = "duckdb",
  catalog = NULL,
  namespace = "ol",
  object_mode = c("blobs", "external"),
  object_root = NULL
)
}
\description{
Initialize DuckDB backend (internal)
}
