% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{.ol_is_adapter_object}
\alias{.ol_is_adapter_object}
\title{Check if object is adapter-managed}
\usage{
.ol_is_adapter_object(state, name, ref = "@latest")
}
\description{
Check if object is adapter-managed
}
\keyword{internal}
