% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_00_serialized_base.R
\name{.ol_known_adapter_tokens}
\alias{.ol_known_adapter_tokens}
\title{Serialized Adapter Helpers}
\usage{
.ol_known_adapter_tokens()
}
\description{
Internal helper factory for object-serialization adapters.
}
\keyword{internal}
