% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{.ol_lineage_to_igraph}
\alias{.ol_lineage_to_igraph}
\title{Convert lineage data.frame to igraph object}
\usage{
.ol_lineage_to_igraph(lineage, focal_node)
}
\arguments{
\item{lineage}{Data.frame with parent, child, type, relationship columns}

\item{focal_node}{Name of the focal node (will be highlighted)}
}
\value{
An igraph object
}
\description{
Convert lineage data.frame to igraph object
}
\keyword{internal}
