% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{.ol_merge_config}
\alias{.ol_merge_config}
\title{Merge two configuration lists (recursive)}
\usage{
.ol_merge_config(defaults, override)
}
\description{
Merge two configuration lists (recursive)
}
\keyword{internal}
