% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{.ol_migrate_dependencies_table}
\alias{.ol_migrate_dependencies_table}
\title{Migrate dependencies table to add version columns}
\usage{
.ol_migrate_dependencies_table(state)
}
\description{
Migrate dependencies table to add version columns
}
\keyword{internal}
