% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\docType{data}
\name{.ol_migration_guide_lines}
\alias{.ol_migration_guide_lines}
\title{Show migration guide for ol_* to Lake API}
\format{
An object of class \code{character} of length 116.
}
\usage{
.ol_migration_guide_lines
}
\description{
Show migration guide for ol_* to Lake API
}
\keyword{datasets}
