% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{ends_with_str}
\alias{ends_with_str}
\title{Check if string ends with a suffix}
\usage{
ends_with_str(x, suffix)
}
\arguments{
\item{x}{Character vector}

\item{suffix}{Suffix to check for}
}
\value{
Logical vector
}
\description{
Check if string ends with a suffix
}
