% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_epigenomics.R
\name{EpigenomicsAdapter}
\alias{EpigenomicsAdapter}
\title{Epigenomics Adapter}
\description{
Adapter for storing and retrieving epigenomics-layer objects.
}
\details{
This umbrella adapter is lower priority than specific adapters
(e.g. Methylation / ATAC / ChIP) and captures epigenomics-marked objects.
}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{EpigenomicsAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EpigenomicsAdapter-name}{\code{EpigenomicsAdapter$name()}}
\item \href{#method-EpigenomicsAdapter-priority}{\code{EpigenomicsAdapter$priority()}}
\item \href{#method-EpigenomicsAdapter-put}{\code{EpigenomicsAdapter$put()}}
\item \href{#method-EpigenomicsAdapter-get}{\code{EpigenomicsAdapter$get()}}
\item \href{#method-EpigenomicsAdapter-components}{\code{EpigenomicsAdapter$components()}}
\item \href{#method-EpigenomicsAdapter-exists}{\code{EpigenomicsAdapter$exists()}}
\item \href{#method-EpigenomicsAdapter-list_names}{\code{EpigenomicsAdapter$list_names()}}
\item \href{#method-EpigenomicsAdapter-can_handle}{\code{EpigenomicsAdapter$can_handle()}}
\item \href{#method-EpigenomicsAdapter-clone}{\code{EpigenomicsAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EpigenomicsAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-EpigenomicsAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EpigenomicsAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EpigenomicsAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-EpigenomicsAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EpigenomicsAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EpigenomicsAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-EpigenomicsAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EpigenomicsAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EpigenomicsAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-EpigenomicsAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EpigenomicsAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EpigenomicsAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-EpigenomicsAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EpigenomicsAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EpigenomicsAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-EpigenomicsAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EpigenomicsAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EpigenomicsAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-EpigenomicsAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EpigenomicsAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EpigenomicsAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-EpigenomicsAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EpigenomicsAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EpigenomicsAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EpigenomicsAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EpigenomicsAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
