% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_case_study.R
\name{eval_case_study}
\alias{eval_case_study}
\title{OmicsLake Case Study (W3)}
\value{
`NULL`. This topic documents internal evaluation helpers.
}
\description{
RNA-seq case study demonstrating reproducibility features
}
