% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_generate.R
\name{eval_generate}
\alias{eval_generate}
\title{OmicsLake Evaluation Data Generators}
\value{
`NULL`. This topic documents internal evaluation helpers.
}
\description{
Functions for generating synthetic datasets for benchmarking
}
