% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{fetch}
\alias{fetch}
\title{Read data from the default lake}
\usage{
fetch(name, ...)
}
\arguments{
\item{name}{Name of the data to read}

\item{...}{Additional arguments passed to Lake$get()}
}
\value{
The requested data
}
\description{
Note: This function is named 'fetch' to avoid conflict with base::get()
}
\examples{
use_lake("ex_fetch", root = tempfile())
put("counts", data.frame(gene = c("A", "B"), value = c(1, 2)))
fetch("counts")
}
\seealso{
\code{\link{Lake}}
}
