% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_base.R
\name{find_adapter}
\alias{find_adapter}
\title{Find adapter for data}
\usage{
find_adapter(data)
}
\arguments{
\item{data}{Data to find adapter for}
}
\value{
Matching adapter or NULL
}
\description{
Find adapter for data
}
\keyword{internal}
