% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{from}
\alias{from}
\title{Start a query from a table on the default lake}
\usage{
from(table)
}
\arguments{
\item{table}{Table name}
}
\value{
A QueryBuilder instance
}
\description{
Start a query from a table on the default lake
}
\examples{
use_lake("ex_from", root = tempfile())
put("t", data.frame(x = 1:3))
qb <- from("t")
}
