% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_genomics.R
\name{GenomicsAdapter}
\alias{GenomicsAdapter}
\title{Genomics Adapter}
\value{
An R6 generator for a \code{LakeAdapter} subclass that
  serializes and restores objects of this omics layer.
}
\description{
Adapter for storing and retrieving genomics-layer objects.
}
\details{
This umbrella adapter is lower priority than specific adapters
(e.g. VCF / RaggedExperiment) and captures genomics-marked objects.
}
\examples{
adapter <- GenomicsAdapter$new()
class(adapter)
}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{GenomicsAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GenomicsAdapter-name}{\code{GenomicsAdapter$name()}}
\item \href{#method-GenomicsAdapter-priority}{\code{GenomicsAdapter$priority()}}
\item \href{#method-GenomicsAdapter-put}{\code{GenomicsAdapter$put()}}
\item \href{#method-GenomicsAdapter-get}{\code{GenomicsAdapter$get()}}
\item \href{#method-GenomicsAdapter-components}{\code{GenomicsAdapter$components()}}
\item \href{#method-GenomicsAdapter-exists}{\code{GenomicsAdapter$exists()}}
\item \href{#method-GenomicsAdapter-list_names}{\code{GenomicsAdapter$list_names()}}
\item \href{#method-GenomicsAdapter-can_handle}{\code{GenomicsAdapter$can_handle()}}
\item \href{#method-GenomicsAdapter-clone}{\code{GenomicsAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenomicsAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-GenomicsAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomicsAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenomicsAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-GenomicsAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomicsAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenomicsAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-GenomicsAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomicsAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenomicsAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-GenomicsAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomicsAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenomicsAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-GenomicsAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomicsAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenomicsAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-GenomicsAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomicsAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenomicsAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-GenomicsAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomicsAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenomicsAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-GenomicsAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomicsAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenomicsAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GenomicsAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomicsAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
