% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_glycomics.R
\name{GlycomicsAdapter}
\alias{GlycomicsAdapter}
\title{Glycomics Adapter}
\description{
Adapter for storing and retrieving glycomics-layer objects.
}
\details{
Supports explicit glycomics marker classes/metadata.
}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{GlycomicsAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GlycomicsAdapter-name}{\code{GlycomicsAdapter$name()}}
\item \href{#method-GlycomicsAdapter-priority}{\code{GlycomicsAdapter$priority()}}
\item \href{#method-GlycomicsAdapter-put}{\code{GlycomicsAdapter$put()}}
\item \href{#method-GlycomicsAdapter-get}{\code{GlycomicsAdapter$get()}}
\item \href{#method-GlycomicsAdapter-components}{\code{GlycomicsAdapter$components()}}
\item \href{#method-GlycomicsAdapter-exists}{\code{GlycomicsAdapter$exists()}}
\item \href{#method-GlycomicsAdapter-list_names}{\code{GlycomicsAdapter$list_names()}}
\item \href{#method-GlycomicsAdapter-can_handle}{\code{GlycomicsAdapter$can_handle()}}
\item \href{#method-GlycomicsAdapter-clone}{\code{GlycomicsAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlycomicsAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-GlycomicsAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlycomicsAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlycomicsAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-GlycomicsAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlycomicsAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlycomicsAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-GlycomicsAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlycomicsAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlycomicsAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-GlycomicsAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlycomicsAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlycomicsAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-GlycomicsAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlycomicsAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlycomicsAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-GlycomicsAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlycomicsAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlycomicsAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-GlycomicsAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlycomicsAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlycomicsAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-GlycomicsAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlycomicsAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlycomicsAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GlycomicsAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlycomicsAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
