% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%between\%}
\alias{\%between\%}
\title{BETWEEN operator for range filtering}
\usage{
x \%between\% range
}
\arguments{
\item{x}{Numeric vector}

\item{range}{Numeric vector of length 2 (min, max), inclusive}
}
\value{
Logical vector
}
\description{
BETWEEN operator for range filtering
}
\examples{
values <- 1:20
values[values \%between\% c(5, 15)]
}
