% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_compat.R
\name{\%>>\%}
\alias{\%>>\%}
\title{Pipe operator that saves to lake}
\usage{
.data \%>>\% target
}
\arguments{
\item{.data}{Data from pipe}

\item{target}{Target specification ("name" or "project/name")}
}
\value{
Invisibly returns the data
}
\description{
Use this operator at the end of a pipe to save results to a lake.
Format: data %>>% "lake_name" or data %>>% "project/name"
}
\examples{
if (FALSE) {
use_lake("my_project")

df |>
  dplyr::filter(x > 5) \%>>\%
  "filtered_data"
}
}
