% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%iregex\%}
\alias{\%iregex\%}
\title{Case-insensitive regex match}
\usage{
x \%iregex\% pattern
}
\arguments{
\item{x}{Character vector}

\item{pattern}{Regular expression pattern}
}
\value{
Logical vector
}
\description{
Case-insensitive regex match
}
