% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_case_study.R, R/eval_metrics.R,
%   R/eval_plot.R
\name{\%||\%}
\alias{\%||\%}
\title{Null coalescing (internal)}
\usage{
x \%||\% y

x \%||\% y

x \%||\% y
}
\description{
Null coalescing (internal)

Null coalescing operator

Null coalescing (internal)
}
\keyword{internal}
