% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%regex\%}
\alias{\%regex\%}
\title{Regex match operator}
\usage{
x \%regex\% pattern
}
\arguments{
\item{x}{Character vector}

\item{pattern}{Regular expression pattern}
}
\value{
Logical vector
}
\description{
Regex match operator
}
\examples{
genes <- c("ENSG00000001", "ENSG00000002", "MT-CO1")
genes[genes \%regex\% "^ENSG"]
}
