% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{history}
\alias{history}
\title{Show history from the default lake}
\usage{
history(name = NULL, ...)
}
\arguments{
\item{name}{Optional data name}

\item{...}{Additional arguments passed to Lake$history()}
}
\value{
History data frame
}
\description{
Show history from the default lake
}
\examples{
use_lake("ex_history", root = tempfile())
put("t", data.frame(x = 1:3))
history()
}
