% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{if_else_na}
\alias{if_else_na}
\title{If-else with NA handling}
\usage{
if_else_na(condition, true, false, na = NA)
}
\arguments{
\item{condition}{Logical vector}

\item{true}{Value when TRUE}

\item{false}{Value when FALSE}

\item{na}{Value when NA (default: NA)}
}
\value{
Vector with conditional values
}
\description{
If-else with NA handling
}
