% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{import_data}
\alias{import_data}
\title{Import data into the default lake}
\usage{
import_data(path, name, ...)
}
\arguments{
\item{path}{File path to import}

\item{name}{Name to store as}

\item{...}{Additional arguments}
}
\value{
Invisible Lake object
}
\description{
Import data into the default lake
}
\examples{
use_lake("ex_import", root = tempfile())
f <- file.path(tempdir(), "imp.csv")
write.csv(data.frame(x = 1:3), f, row.names = FALSE)
import_data(f, "imported")
}
