% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_compat.R
\name{into}
\alias{into}
\title{Create a pipe-compatible lake write function}
\usage{
into(lake)
}
\arguments{
\item{lake}{Lake instance}
}
\value{
A function that accepts data and name, saves to lake
}
\description{
Creates a function that can be used at the end of a pipe to save data.
}
\examples{
if (FALSE) {
write_to <- into(lake)

df |>
  dplyr::filter(x > 5) |>
  write_to("filtered_data")
}
}
