% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{is_not_null}
\alias{is_not_null}
\title{Check for non-NULL/non-NA values}
\usage{
is_not_null(x)
}
\arguments{
\item{x}{Vector to check}
}
\value{
Logical vector indicating non-NA values
}
\description{
Check for non-NULL/non-NA values
}
