% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{is_null}
\alias{is_null}
\title{Check for NULL/NA values}
\usage{
is_null(x)
}
\arguments{
\item{x}{Vector to check}
}
\value{
Logical vector indicating NA values
}
\description{
Check for NULL/NA values
}
\examples{
x <- c(1, NA, 3, NA, 5)
x[is_null(x)]
}
