% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{lake_aliases}
\alias{lake_aliases}
\alias{lake_use}
\alias{lake_put}
\alias{lake_get}
\alias{lake_ref}
\alias{lake_snap}
\alias{lake_tag}
\alias{lake_tree}
\alias{lake_has}
\alias{lake_track}
\alias{lake_track_script}
\alias{lake_auto_on}
\alias{lake_auto_off}
\alias{lake_strict_on}
\title{Simple `lake_*` Aliases}
\usage{
lake_use(project = NULL, ...)

lake_put(name, data, ...)

lake_get(name, ...)

lake_ref(name)

lake_snap(label, ...)

lake_tag(name, tag)

lake_tree(name = NULL, ...)

lake_has(name, type = c("any", "table", "object"))

lake_track(expr, ...)

lake_track_script(path, ...)

lake_auto_on(...)

lake_auto_off(commit = TRUE)

lake_strict_on(...)
}
\arguments{
\item{project}{Project name for the default lake}

\item{...}{Additional arguments forwarded to the corresponding function}

\item{name}{Data/table/node name (meaning depends on each function)}

\item{data}{Data to store}

\item{label}{Snapshot label}

\item{tag}{Tag name}

\item{type}{One of "any", "table", or "object"}

\item{expr}{Expression to track}

\item{path}{Script file path}

\item{commit}{If TRUE, write tracked lineage before disabling transparent
mode}
}
\description{
Consistent, memorable aliases that follow one naming rule:
use `lake_<verb>` for common operations.
}
