% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{lake_exists}
\alias{lake_exists}
\title{Check whether a data name exists in the default lake}
\usage{
lake_exists(name, type = c("any", "table", "object"))
}
\arguments{
\item{name}{Data name}

\item{type}{One of "any", "table", or "object"}
}
\value{
TRUE/FALSE
}
\description{
Check whether a data name exists in the default lake
}
