% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods_docs.R
\name{lake_tbl_methods}
\alias{lake_tbl_methods}
\alias{filter.lake_tbl}
\alias{select.lake_tbl}
\alias{mutate.lake_tbl}
\alias{summarise.lake_tbl}
\alias{summarize.lake_tbl}
\alias{group_by.lake_tbl}
\alias{ungroup.lake_tbl}
\alias{arrange.lake_tbl}
\alias{distinct.lake_tbl}
\alias{slice.lake_tbl}
\alias{rename.lake_tbl}
\alias{relocate.lake_tbl}
\alias{left_join.lake_tbl}
\alias{inner_join.lake_tbl}
\alias{right_join.lake_tbl}
\alias{full_join.lake_tbl}
\alias{semi_join.lake_tbl}
\alias{anti_join.lake_tbl}
\alias{collect.lake_tbl}
\title{lake_tbl S3 Methods}
\description{
Internal S3 methods for dplyr verbs on `lake_tbl` objects. These methods
preserve lineage metadata through table operations.
}
\keyword{internal}
