% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_lipidomics.R
\name{LipidomicsAdapter}
\alias{LipidomicsAdapter}
\title{Lipidomics Adapter}
\value{
An R6 generator for a \code{LakeAdapter} subclass that
  serializes and restores objects of this omics layer.
}
\description{
Adapter for storing and retrieving lipidomics-layer objects.
}
\details{
Supports explicit lipidomics marker classes/metadata.
}
\examples{
adapter <- LipidomicsAdapter$new()
class(adapter)
}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{LipidomicsAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LipidomicsAdapter-name}{\code{LipidomicsAdapter$name()}}
\item \href{#method-LipidomicsAdapter-priority}{\code{LipidomicsAdapter$priority()}}
\item \href{#method-LipidomicsAdapter-put}{\code{LipidomicsAdapter$put()}}
\item \href{#method-LipidomicsAdapter-get}{\code{LipidomicsAdapter$get()}}
\item \href{#method-LipidomicsAdapter-components}{\code{LipidomicsAdapter$components()}}
\item \href{#method-LipidomicsAdapter-exists}{\code{LipidomicsAdapter$exists()}}
\item \href{#method-LipidomicsAdapter-list_names}{\code{LipidomicsAdapter$list_names()}}
\item \href{#method-LipidomicsAdapter-can_handle}{\code{LipidomicsAdapter$can_handle()}}
\item \href{#method-LipidomicsAdapter-clone}{\code{LipidomicsAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LipidomicsAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-LipidomicsAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LipidomicsAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LipidomicsAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-LipidomicsAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LipidomicsAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LipidomicsAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-LipidomicsAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LipidomicsAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LipidomicsAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-LipidomicsAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LipidomicsAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LipidomicsAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-LipidomicsAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LipidomicsAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LipidomicsAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-LipidomicsAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LipidomicsAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LipidomicsAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-LipidomicsAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LipidomicsAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LipidomicsAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-LipidomicsAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LipidomicsAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LipidomicsAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LipidomicsAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LipidomicsAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
