% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_mae.R
\name{MAEAdapter}
\alias{MAEAdapter}
\title{MultiAssayExperiment Adapter}
\description{
Adapter for storing and retrieving
\code{MultiAssayExperiment} objects.
}
\details{
This adapter preserves \code{MultiAssayExperiment} components:
\itemize{
  \item experiments (\code{experiments})
  \item sample map (\code{sampleMap})
  \item participant metadata (\code{colData})
  \item dropped samples (\code{drops})
  \item object metadata (\code{metadata})
}
}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{MAEAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MAEAdapter-name}{\code{MAEAdapter$name()}}
\item \href{#method-MAEAdapter-can_handle}{\code{MAEAdapter$can_handle()}}
\item \href{#method-MAEAdapter-priority}{\code{MAEAdapter$priority()}}
\item \href{#method-MAEAdapter-put}{\code{MAEAdapter$put()}}
\item \href{#method-MAEAdapter-get}{\code{MAEAdapter$get()}}
\item \href{#method-MAEAdapter-components}{\code{MAEAdapter$components()}}
\item \href{#method-MAEAdapter-exists}{\code{MAEAdapter$exists()}}
\item \href{#method-MAEAdapter-list_names}{\code{MAEAdapter$list_names()}}
\item \href{#method-MAEAdapter-clone}{\code{MAEAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MAEAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-MAEAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAEAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MAEAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-MAEAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAEAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MAEAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-MAEAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAEAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MAEAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-MAEAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAEAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MAEAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-MAEAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAEAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MAEAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-MAEAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAEAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MAEAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-MAEAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAEAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MAEAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-MAEAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAEAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MAEAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MAEAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAEAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
