% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_metabolomics.R
\name{MetabolomicsAdapter}
\alias{MetabolomicsAdapter}
\title{Metabolomics Adapter}
\value{
An R6 generator for a \code{LakeAdapter} subclass that
  serializes and restores objects of this omics layer.
}
\description{
Adapter for storing and retrieving metabolomics-layer objects.
}
\details{
Supports explicit metabolomics marker classes/metadata.
}
\examples{
adapter <- MetabolomicsAdapter$new()
class(adapter)
}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{MetabolomicsAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MetabolomicsAdapter-name}{\code{MetabolomicsAdapter$name()}}
\item \href{#method-MetabolomicsAdapter-priority}{\code{MetabolomicsAdapter$priority()}}
\item \href{#method-MetabolomicsAdapter-put}{\code{MetabolomicsAdapter$put()}}
\item \href{#method-MetabolomicsAdapter-get}{\code{MetabolomicsAdapter$get()}}
\item \href{#method-MetabolomicsAdapter-components}{\code{MetabolomicsAdapter$components()}}
\item \href{#method-MetabolomicsAdapter-exists}{\code{MetabolomicsAdapter$exists()}}
\item \href{#method-MetabolomicsAdapter-list_names}{\code{MetabolomicsAdapter$list_names()}}
\item \href{#method-MetabolomicsAdapter-can_handle}{\code{MetabolomicsAdapter$can_handle()}}
\item \href{#method-MetabolomicsAdapter-clone}{\code{MetabolomicsAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetabolomicsAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-MetabolomicsAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetabolomicsAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetabolomicsAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-MetabolomicsAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetabolomicsAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetabolomicsAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-MetabolomicsAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetabolomicsAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetabolomicsAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-MetabolomicsAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetabolomicsAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetabolomicsAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-MetabolomicsAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetabolomicsAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetabolomicsAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-MetabolomicsAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetabolomicsAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetabolomicsAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-MetabolomicsAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetabolomicsAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetabolomicsAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-MetabolomicsAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetabolomicsAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetabolomicsAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MetabolomicsAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetabolomicsAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
