% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_methylation.R
\name{MethylationAdapter}
\alias{MethylationAdapter}
\title{Methylation Adapter}
\description{
Adapter for storing and retrieving methylation-layer objects.
}
\details{
Supports common Bioconductor methylation classes (e.g. minfi family) by
storing a full serialized object and manifest.
}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{MethylationAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MethylationAdapter-name}{\code{MethylationAdapter$name()}}
\item \href{#method-MethylationAdapter-priority}{\code{MethylationAdapter$priority()}}
\item \href{#method-MethylationAdapter-put}{\code{MethylationAdapter$put()}}
\item \href{#method-MethylationAdapter-get}{\code{MethylationAdapter$get()}}
\item \href{#method-MethylationAdapter-components}{\code{MethylationAdapter$components()}}
\item \href{#method-MethylationAdapter-exists}{\code{MethylationAdapter$exists()}}
\item \href{#method-MethylationAdapter-list_names}{\code{MethylationAdapter$list_names()}}
\item \href{#method-MethylationAdapter-can_handle}{\code{MethylationAdapter$can_handle()}}
\item \href{#method-MethylationAdapter-clone}{\code{MethylationAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MethylationAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-MethylationAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MethylationAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MethylationAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-MethylationAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MethylationAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MethylationAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-MethylationAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MethylationAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MethylationAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-MethylationAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MethylationAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MethylationAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-MethylationAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MethylationAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MethylationAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-MethylationAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MethylationAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MethylationAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-MethylationAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MethylationAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MethylationAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-MethylationAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MethylationAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MethylationAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MethylationAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MethylationAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
