% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_msexperiment.R
\name{MsExperimentAdapter}
\alias{MsExperimentAdapter}
\title{MsExperiment Adapter}
\description{
Adapter for storing and retrieving
\code{MsExperiment} objects.
}
\details{
This adapter preserves \code{MsExperiment} components:
\itemize{
  \item sample metadata (\code{sampleData})
  \item spectra (\code{spectra})
  \item quantitative/assay data (\code{qdata})
  \item experiment files (\code{experimentFiles})
  \item auxiliary data (\code{otherData})
}
}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{MsExperimentAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MsExperimentAdapter-name}{\code{MsExperimentAdapter$name()}}
\item \href{#method-MsExperimentAdapter-can_handle}{\code{MsExperimentAdapter$can_handle()}}
\item \href{#method-MsExperimentAdapter-priority}{\code{MsExperimentAdapter$priority()}}
\item \href{#method-MsExperimentAdapter-put}{\code{MsExperimentAdapter$put()}}
\item \href{#method-MsExperimentAdapter-get}{\code{MsExperimentAdapter$get()}}
\item \href{#method-MsExperimentAdapter-components}{\code{MsExperimentAdapter$components()}}
\item \href{#method-MsExperimentAdapter-exists}{\code{MsExperimentAdapter$exists()}}
\item \href{#method-MsExperimentAdapter-list_names}{\code{MsExperimentAdapter$list_names()}}
\item \href{#method-MsExperimentAdapter-clone}{\code{MsExperimentAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsExperimentAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-MsExperimentAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsExperimentAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsExperimentAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-MsExperimentAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsExperimentAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsExperimentAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-MsExperimentAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsExperimentAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsExperimentAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-MsExperimentAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsExperimentAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsExperimentAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-MsExperimentAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsExperimentAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsExperimentAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-MsExperimentAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsExperimentAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsExperimentAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-MsExperimentAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsExperimentAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsExperimentAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-MsExperimentAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsExperimentAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsExperimentAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MsExperimentAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsExperimentAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
