% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{not_between_operator}
\alias{not_between_operator}
\alias{\%!between\%}
\title{NOT BETWEEN operator}
\usage{
x \%!between\% range
}
\arguments{
\item{x}{Numeric vector}

\item{range}{Numeric vector of length 2 (min, max)}
}
\value{
Logical vector
}
\description{
NOT BETWEEN operator
}
