% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{not_in_operator}
\alias{not_in_operator}
\alias{\%!in\%}
\title{NOT IN operator}
\usage{
x \%!in\% table
}
\arguments{
\item{x}{Vector}

\item{table}{Values to exclude}
}
\value{
Logical vector
}
\description{
NOT IN operator
}
\examples{
letters[letters \%!in\% c("a", "e", "i", "o", "u")]
}
