% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{objects}
\alias{objects}
\title{List objects in the default lake}
\usage{
objects()
}
\value{
Data frame of object names
}
\description{
List objects in the default lake
}
\examples{
use_lake("ex_objects", root = tempfile())
put("model", list(a = 1, b = 2))
objects()
}
