% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe.R
\name{observe_session}
\alias{observe_session}
\title{Create an observed session}
\usage{
observe_session(lake)
}
\arguments{
\item{lake}{Lake instance to observe}
}
\value{
An ObserveSession object
}
\description{
Starts an observation session that tracks all Lake operations until stopped.
}
\examples{
if (FALSE) {
    lake <- Lake$new("project")
    session <- observe_session(lake)

    # Do work...
    lake$put("data1", df1)
    lake$put("data2", df2, depends_on = "data1")

    # Stop and get summary
    summary <- session$stop()
    print(summary)
}
}
