% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe.R
\name{ObserveSession}
\alias{ObserveSession}
\title{Observation Session}
\description{
Tracks Lake operations over time
}
\keyword{Check}
\keyword{Record}
\keyword{Stop}
\keyword{active}
\keyword{an}
\keyword{if}
\keyword{internal}
\keyword{is}
\keyword{operation}
\keyword{session}
\keyword{the}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ObserveSession-new}{\code{ObserveSession$new()}}
\item \href{#method-ObserveSession-record}{\code{ObserveSession$record()}}
\item \href{#method-ObserveSession-stop}{\code{ObserveSession$stop()}}
\item \href{#method-ObserveSession-is_active}{\code{ObserveSession$is_active()}}
\item \href{#method-ObserveSession-print}{\code{ObserveSession$print()}}
\item \href{#method-ObserveSession-clone}{\code{ObserveSession$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObserveSession-new"></a>}}
\if{latex}{\out{\hypertarget{method-ObserveSession-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObserveSession$new(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObserveSession-record"></a>}}
\if{latex}{\out{\hypertarget{method-ObserveSession-record}{}}}
\subsection{Method \code{record()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObserveSession$record(op_type, name, details = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObserveSession-stop"></a>}}
\if{latex}{\out{\hypertarget{method-ObserveSession-stop}{}}}
\subsection{Method \code{stop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObserveSession$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObserveSession-is_active"></a>}}
\if{latex}{\out{\hypertarget{method-ObserveSession-is_active}{}}}
\subsection{Method \code{is_active()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObserveSession$is_active()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObserveSession-print"></a>}}
\if{latex}{\out{\hypertarget{method-ObserveSession-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObserveSession$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObserveSession-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ObserveSession-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObserveSession$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
