% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{ol_cumulative_sum}
\alias{ol_cumulative_sum}
\title{Calculate cumulative sum}
\usage{
ol_cumulative_sum(
  table,
  column,
  partition_by = NULL,
  order_by,
  as_column = NULL,
  project = getOption("ol.project"),
  collect = TRUE
)
}
\arguments{
\item{table}{Character string, name of the table}

\item{column}{Character string, column name to calculate cumulative sum on}

\item{partition_by}{Character vector of columns to partition by (default:
NULL)}

\item{order_by}{Character string, column to order by (required)}

\item{as_column}{Character, name of the output column (default:
paste0(column, "_cumsum"))}

\item{project}{Project name (default: current project)}

\item{collect}{Logical; if TRUE returns data.frame, if FALSE returns lazy
dplyr table}
}
\value{
Original table with added cumulative sum column
}
\description{
Add a cumulative sum column using window functions.
}
\examples{
if (FALSE) {
    ol_cumulative_sum("counts", "value", order_by = "time")

    ol_cumulative_sum("counts", "value",
        partition_by = "sample",
        order_by = "gene_id"
    )
}
}
