% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{ol_drop_view}
\alias{ol_drop_view}
\title{Drop a database view}
\usage{
ol_drop_view(name, project = getOption("ol.project"))
}
\arguments{
\item{name}{Name of the view to drop}

\item{project}{Project name (default: current project from options)}
}
\value{
Invisible TRUE if successful
}
\description{
Drops a view from the project. Also removes any dependency records
associated with the view.
}
\examples{
if (FALSE) {
    ol_init("myproject")
    ol_create_view("my_view", "SELECT * FROM genes WHERE expr > 0")

    ol_drop_view("my_view")
}
}
