% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{ol_eval_bench}
\alias{ol_eval_bench}
\title{Measure with multiple repetitions}
\usage{
ol_eval_bench(expr, n = 10, warmup = 1, measure_memory = FALSE)
}
\arguments{
\item{expr}{Expression to evaluate}

\item{n}{Number of repetitions}

\item{warmup}{Number of warmup runs (not recorded)}

\item{measure_memory}{Whether to measure memory}
}
\value{
List with times, median, iqr, min, max
}
\description{
Measure with multiple repetitions
}
\keyword{internal}
