% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{ol_eval_capture_packages}
\alias{ol_eval_capture_packages}
\title{Capture package versions for key dependencies}
\usage{
ol_eval_capture_packages()
}
\value{
Named list of package versions
}
\description{
Capture package versions for key dependencies
}
\keyword{internal}
