% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_case_study.R
\name{ol_eval_case_study_report}
\alias{ol_eval_case_study_report}
\title{Generate case study report}
\usage{
ol_eval_case_study_report(results, output_file = NULL)
}
\arguments{
\item{results}{Results from ol_eval_run_case_study()}

\item{output_file}{Output markdown file path}
}
\value{
Invisible path to output file
}
\description{
Generate case study report
}
\keyword{internal}
