% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_generate.R
\name{ol_eval_generate_object}
\alias{ol_eval_generate_object}
\title{Generate a list object for object storage benchmarks}
\usage{
ol_eval_generate_object(size = c("small", "medium", "large"), seed = 1)
}
\arguments{
\item{size}{Approximate target size: "small", "medium", or "large"}

\item{seed}{Random seed}
}
\value{
A complex list suitable for object serialization benchmarks
}
\description{
Generate a list object for object storage benchmarks
}
\keyword{internal}
