% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{ol_eval_get_threads}
\alias{ol_eval_get_threads}
\title{Get number of threads available}
\usage{
ol_eval_get_threads()
}
\value{
Integer number of threads
}
\description{
Get number of threads available
}
\keyword{internal}
