% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_plot.R
\name{ol_eval_plot_lineage}
\alias{ol_eval_plot_lineage}
\title{Plot lineage operation scaling}
\usage{
ol_eval_plot_lineage(df, output_dir, format = "png")
}
\arguments{
\item{df}{Results data frame}

\item{output_dir}{Output directory}

\item{format}{Output format}
}
\value{
Path to saved plot
}
\description{
Plot lineage operation scaling
}
\keyword{internal}
