% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_plot.R
\name{ol_eval_plot_storage}
\alias{ol_eval_plot_storage}
\title{Plot storage growth with tags/snaps}
\usage{
ol_eval_plot_storage(df, output_dir, format = "png")
}
\arguments{
\item{df}{Results data frame}

\item{output_dir}{Output directory}

\item{format}{Output format}
}
\value{
Path to saved plot
}
\description{
Plot storage growth with tags/snaps
}
\keyword{internal}
