% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{ol_eval_read_jsonl}
\alias{ol_eval_read_jsonl}
\title{Read results from JSONL file}
\usage{
ol_eval_read_jsonl(file)
}
\arguments{
\item{file}{Path to JSONL file}
}
\value{
List of result records
}
\description{
Read results from JSONL file
}
\keyword{internal}
