% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ol_iceberg_adapter.R
\name{ol_init_iceberg}
\alias{ol_init_iceberg}
\title{Initialize DuckDB + Iceberg backend}
\usage{
ol_init_iceberg(
  project,
  engine = "duckdb",
  catalog = NULL,
  namespace = "ol",
  object_mode = c("blobs", "external"),
  object_root = NULL
)
}
\description{
Initialize DuckDB + Iceberg backend
}
