% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ol_list_tags}
\alias{ol_list_tags}
\title{List all tags for a table or all project labels}
\usage{
ol_list_tags(name = NULL, project = getOption("ol.project"))
}
\arguments{
\item{name}{Optional table name to list tags for. If NULL, lists all tags.}

\item{project}{Project name}
}
\value{
Data frame of tags
}
\description{
List all tags for a table or all project labels
}
\examples{
ol_init("ex_ol_list_tags", root = tempfile())
ol_write("t", data.frame(x = 1:3))
ol_list_tags()
}
