% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ol_load}
\alias{ol_load}
\title{Alias of \code{ol_read()}}
\usage{
ol_load(name, ref = "@latest", project = getOption("ol.project"))
}
\arguments{
\item{name}{Table or object name}

\item{ref}{Reference string (e.g. "@latest", "@tag(v1)")}

\item{project}{Project name}
}
\value{
Data frame, lazy table, or stored object
}
\description{
Alias of \code{ol_read()}
}
