% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ol_log_commits}
\alias{ol_log_commits}
\title{View commit history}
\usage{
ol_log_commits(project = getOption("ol.project"), n = 20)
}
\arguments{
\item{project}{Project name}

\item{n}{Maximum number of commits to return}
}
\value{
Data frame of commits with reproducibility and agent metadata columns
}
\description{
View commit history
}
\examples{
ol_init("ex_ol_log_commits", root = tempfile())
ol_write("t", data.frame(x = 1:3))
ol_commit("note")
ol_log_commits()
}
