% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ol_save}
\alias{ol_save}
\title{Save an R object via the backend metadata table}
\usage{
ol_save(
  name,
  object,
  project = getOption("ol.project"),
  mime = NULL,
  depends_on = NULL
)
}
\arguments{
\item{name}{Object name}

\item{object}{R object to save}

\item{project}{Project name}

\item{mime}{MIME type for object payload}

\item{depends_on}{Optional character vector of table/object names that this object depends on}
}
\value{
Invisible TRUE on success
}
\description{
Save an R object via the backend metadata table
}
