% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{OmicsLake-package}
\alias{OmicsLake}
\alias{OmicsLake-package}
\title{OmicsLake: OmicsLake: Versioned, On-Disk Omics Data Management for Bioconductor}
\description{
A lightweight framework for versioned, on-disk omics data management using DuckDB + Arrow + Parquet. Provides simple read/write-style functions and Bioconductor-compatible readers (SummarizedExperiment, MultiAssayExperiment).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matsui-lab/OmicsLake}
  \item Report bugs at \url{https://github.com/matsui-lab/OmicsLake/issues}
}

}
\author{
\strong{Maintainer}: Yusuke Matsui \email{mail.to.matsui@gmail.com} (\href{https://orcid.org/0000-0003-3977-4313}{ORCID})

}
