% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_phosphoproteomics.R
\name{PhosphoproteomicsAdapter}
\alias{PhosphoproteomicsAdapter}
\title{Phosphoproteomics Adapter}
\value{
An R6 generator for a \code{LakeAdapter} subclass that
  serializes and restores objects of this omics layer.
}
\description{
Adapter for storing and retrieving phosphoproteomics-layer
objects.
}
\details{
Supports explicit phosphoproteomics marker classes/metadata.
}
\examples{
adapter <- PhosphoproteomicsAdapter$new()
class(adapter)
}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{PhosphoproteomicsAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PhosphoproteomicsAdapter-name}{\code{PhosphoproteomicsAdapter$name()}}
\item \href{#method-PhosphoproteomicsAdapter-priority}{\code{PhosphoproteomicsAdapter$priority()}}
\item \href{#method-PhosphoproteomicsAdapter-put}{\code{PhosphoproteomicsAdapter$put()}}
\item \href{#method-PhosphoproteomicsAdapter-get}{\code{PhosphoproteomicsAdapter$get()}}
\item \href{#method-PhosphoproteomicsAdapter-components}{\code{PhosphoproteomicsAdapter$components()}}
\item \href{#method-PhosphoproteomicsAdapter-exists}{\code{PhosphoproteomicsAdapter$exists()}}
\item \href{#method-PhosphoproteomicsAdapter-list_names}{\code{PhosphoproteomicsAdapter$list_names()}}
\item \href{#method-PhosphoproteomicsAdapter-can_handle}{\code{PhosphoproteomicsAdapter$can_handle()}}
\item \href{#method-PhosphoproteomicsAdapter-clone}{\code{PhosphoproteomicsAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhosphoproteomicsAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-PhosphoproteomicsAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhosphoproteomicsAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhosphoproteomicsAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-PhosphoproteomicsAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhosphoproteomicsAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhosphoproteomicsAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-PhosphoproteomicsAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhosphoproteomicsAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhosphoproteomicsAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-PhosphoproteomicsAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhosphoproteomicsAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhosphoproteomicsAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-PhosphoproteomicsAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhosphoproteomicsAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhosphoproteomicsAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-PhosphoproteomicsAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhosphoproteomicsAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhosphoproteomicsAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-PhosphoproteomicsAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhosphoproteomicsAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhosphoproteomicsAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-PhosphoproteomicsAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhosphoproteomicsAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhosphoproteomicsAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PhosphoproteomicsAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhosphoproteomicsAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
