% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe.R
\name{print.lake_observation}
\alias{print.lake_observation}
\title{Print method for lake_observation}
\usage{
\method{print}{lake_observation}(x, ...)
}
\arguments{
\item{x}{A lake_observation object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for lake_observation
}
