% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair.R
\name{print.lake_repair_report}
\alias{print.lake_repair_report}
\title{Print method for lake_repair_report}
\usage{
\method{print}{lake_repair_report}(x, ...)
}
\arguments{
\item{x}{A lake_repair_report object}

\item{...}{Additional arguments (ignored)}
}
\value{
The input object, invisibly
}
\description{
Print method for lake_repair_report
}
