% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_proteomics.R
\name{ProteomicsAdapter}
\alias{ProteomicsAdapter}
\title{Proteomics Adapter}
\value{
An R6 generator for a \code{LakeAdapter} subclass that
  serializes and restores objects of this omics layer.
}
\description{
Adapter for storing and retrieving proteomics-layer objects.
}
\details{
This umbrella adapter is lower priority than specific adapters
(e.g. QFeatures / MsExperiment / Spectra) and captures
proteomics-marked objects.
}
\examples{
adapter <- ProteomicsAdapter$new()
class(adapter)
}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{ProteomicsAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProteomicsAdapter-name}{\code{ProteomicsAdapter$name()}}
\item \href{#method-ProteomicsAdapter-priority}{\code{ProteomicsAdapter$priority()}}
\item \href{#method-ProteomicsAdapter-put}{\code{ProteomicsAdapter$put()}}
\item \href{#method-ProteomicsAdapter-get}{\code{ProteomicsAdapter$get()}}
\item \href{#method-ProteomicsAdapter-components}{\code{ProteomicsAdapter$components()}}
\item \href{#method-ProteomicsAdapter-exists}{\code{ProteomicsAdapter$exists()}}
\item \href{#method-ProteomicsAdapter-list_names}{\code{ProteomicsAdapter$list_names()}}
\item \href{#method-ProteomicsAdapter-can_handle}{\code{ProteomicsAdapter$can_handle()}}
\item \href{#method-ProteomicsAdapter-clone}{\code{ProteomicsAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteomicsAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-ProteomicsAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteomicsAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteomicsAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-ProteomicsAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteomicsAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteomicsAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-ProteomicsAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteomicsAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteomicsAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-ProteomicsAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteomicsAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteomicsAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-ProteomicsAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteomicsAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteomicsAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-ProteomicsAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteomicsAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteomicsAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-ProteomicsAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteomicsAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteomicsAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-ProteomicsAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteomicsAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteomicsAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProteomicsAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteomicsAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
