% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{put}
\alias{put}
\title{Write data to the default lake}
\usage{
put(name, data, ...)
}
\arguments{
\item{name}{Name for the data}

\item{data}{Data to store}

\item{...}{Additional arguments passed to Lake$put()}
}
\value{
Invisible Lake object
}
\description{
Write data to the default lake
}
\examples{
use_lake("ex_put", root = tempfile())
put("counts", data.frame(gene = c("A", "B"), value = c(1, 2)))
}
\seealso{
\code{\link{Lake}}
}
