% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_qfeatures.R
\name{QFeaturesAdapter}
\alias{QFeaturesAdapter}
\title{QFeatures Adapter}
\description{
Adapter for storing and retrieving
\code{QFeatures} objects.
}
\details{
This adapter preserves \code{QFeatures} components:
\itemize{
  \item underlying experiments/sample maps via MultiAssayExperiment adapter
  \item assay links (\code{assayLinks})
}
}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{QFeaturesAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QFeaturesAdapter-name}{\code{QFeaturesAdapter$name()}}
\item \href{#method-QFeaturesAdapter-can_handle}{\code{QFeaturesAdapter$can_handle()}}
\item \href{#method-QFeaturesAdapter-priority}{\code{QFeaturesAdapter$priority()}}
\item \href{#method-QFeaturesAdapter-put}{\code{QFeaturesAdapter$put()}}
\item \href{#method-QFeaturesAdapter-get}{\code{QFeaturesAdapter$get()}}
\item \href{#method-QFeaturesAdapter-components}{\code{QFeaturesAdapter$components()}}
\item \href{#method-QFeaturesAdapter-exists}{\code{QFeaturesAdapter$exists()}}
\item \href{#method-QFeaturesAdapter-list_names}{\code{QFeaturesAdapter$list_names()}}
\item \href{#method-QFeaturesAdapter-clone}{\code{QFeaturesAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QFeaturesAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-QFeaturesAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QFeaturesAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QFeaturesAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-QFeaturesAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QFeaturesAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QFeaturesAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-QFeaturesAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QFeaturesAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QFeaturesAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-QFeaturesAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QFeaturesAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QFeaturesAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-QFeaturesAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QFeaturesAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QFeaturesAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-QFeaturesAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QFeaturesAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QFeaturesAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-QFeaturesAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QFeaturesAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QFeaturesAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-QFeaturesAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QFeaturesAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QFeaturesAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QFeaturesAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QFeaturesAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
