% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{query}
\alias{query}
\title{Start a query builder on the default lake}
\usage{
query()
}
\value{
A QueryBuilder instance
}
\description{
Start a query builder on the default lake
}
\examples{
use_lake("ex_query", root = tempfile())
put("t", data.frame(x = 1:3))
qb <- query()
}
