% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_raggedexperiment.R
\name{RaggedExperimentAdapter}
\alias{RaggedExperimentAdapter}
\title{RaggedExperiment Adapter}
\description{
Adapter for storing and retrieving RaggedExperiment objects.
}
\details{
This adapter stores a full-fidelity serialized RaggedExperiment object and
manifest.
}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{RaggedExperimentAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RaggedExperimentAdapter-name}{\code{RaggedExperimentAdapter$name()}}
\item \href{#method-RaggedExperimentAdapter-priority}{\code{RaggedExperimentAdapter$priority()}}
\item \href{#method-RaggedExperimentAdapter-put}{\code{RaggedExperimentAdapter$put()}}
\item \href{#method-RaggedExperimentAdapter-get}{\code{RaggedExperimentAdapter$get()}}
\item \href{#method-RaggedExperimentAdapter-components}{\code{RaggedExperimentAdapter$components()}}
\item \href{#method-RaggedExperimentAdapter-exists}{\code{RaggedExperimentAdapter$exists()}}
\item \href{#method-RaggedExperimentAdapter-list_names}{\code{RaggedExperimentAdapter$list_names()}}
\item \href{#method-RaggedExperimentAdapter-can_handle}{\code{RaggedExperimentAdapter$can_handle()}}
\item \href{#method-RaggedExperimentAdapter-clone}{\code{RaggedExperimentAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RaggedExperimentAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-RaggedExperimentAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RaggedExperimentAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RaggedExperimentAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-RaggedExperimentAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RaggedExperimentAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RaggedExperimentAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-RaggedExperimentAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RaggedExperimentAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RaggedExperimentAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-RaggedExperimentAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RaggedExperimentAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RaggedExperimentAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-RaggedExperimentAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RaggedExperimentAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RaggedExperimentAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-RaggedExperimentAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RaggedExperimentAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RaggedExperimentAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-RaggedExperimentAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RaggedExperimentAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RaggedExperimentAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-RaggedExperimentAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RaggedExperimentAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RaggedExperimentAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RaggedExperimentAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RaggedExperimentAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
