% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{ref}
\alias{ref}
\title{Get a lazy reference from the default lake}
\usage{
ref(name)
}
\arguments{
\item{name}{Table name}
}
\value{
A lazy table reference for use with dplyr
}
\description{
Get a lazy reference from the default lake
}
\examples{
use_lake("ex_ref", root = tempfile())
put("t", data.frame(x = 1:5))
dplyr::collect(dplyr::filter(ref("t"), x > 2))
}
