% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{restore}
\alias{restore}
\title{Restore the default lake to a snapshot}
\usage{
restore(label)
}
\arguments{
\item{label}{Snapshot label}
}
\value{
Invisible Lake object
}
\description{
Restore the default lake to a snapshot
}
\examples{
use_lake("ex_restore", root = tempfile())
put("t", data.frame(x = 1:3))
snap("v1")
restore("v1")
}
