% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_compat.R
\name{save_as}
\alias{save_as}
\title{Save pipe result to lake}
\usage{
save_as(.data, name, lake = NULL)
}
\arguments{
\item{.data}{Data from pipe (data.frame or tbl_lazy)}

\item{name}{Name to save as in the lake}

\item{lake}{Lake instance. If NULL, uses the default lake from use_lake()}
}
\value{
Invisibly returns the data (for potential further piping)
}
\description{
This function is designed to be used at the end of a dplyr pipe
to save results to a Lake with automatic dependency tracking.
}
\examples{
if (FALSE) {
lake$ref("raw_data") |>
  dplyr::filter(quality > 0.5) |>
  dplyr::mutate(normalized = value / mean(value)) |>
  save_as("processed_data", lake)
}
}
