% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_sce.R
\name{SCEAdapter}
\alias{SCEAdapter}
\title{SingleCellExperiment Adapter}
\description{
Adapter for storing and retrieving
SingleCellExperiment objects.
}
\details{
This adapter extends SummarizedExperiment support by preserving
SingleCellExperiment-specific components:
\itemize{
  \item reduced dimensions (\code{reducedDims})
  \item alternative experiments (\code{altExps})
  \item row/column pairings (\code{rowPairs}, \code{colPairs})
  \item size factors (\code{sizeFactors})
  \item column labels (\code{colLabels})
  \item main experiment label (\code{mainExpName})
}
}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{SCEAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SCEAdapter-name}{\code{SCEAdapter$name()}}
\item \href{#method-SCEAdapter-can_handle}{\code{SCEAdapter$can_handle()}}
\item \href{#method-SCEAdapter-priority}{\code{SCEAdapter$priority()}}
\item \href{#method-SCEAdapter-put}{\code{SCEAdapter$put()}}
\item \href{#method-SCEAdapter-get}{\code{SCEAdapter$get()}}
\item \href{#method-SCEAdapter-components}{\code{SCEAdapter$components()}}
\item \href{#method-SCEAdapter-exists}{\code{SCEAdapter$exists()}}
\item \href{#method-SCEAdapter-list_names}{\code{SCEAdapter$list_names()}}
\item \href{#method-SCEAdapter-clone}{\code{SCEAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCEAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-SCEAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCEAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCEAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-SCEAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCEAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCEAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-SCEAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCEAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCEAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-SCEAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCEAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCEAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-SCEAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCEAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCEAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-SCEAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCEAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCEAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-SCEAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCEAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCEAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-SCEAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCEAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SCEAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SCEAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCEAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
