% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_se.R
\name{SEAdapter}
\alias{SEAdapter}
\title{SummarizedExperiment Adapter}
\description{
Adapter for storing and retrieving Bioconductor
SummarizedExperiment objects.
Stores all components (assays, colData, rowData, metadata) with full
fidelity.
}
\details{
This adapter decomposes a SummarizedExperiment into its components:
\itemize{
  \item Assays are stored as tables (sparse matrix -> long format)
  \item colData is stored as a table
  \item rowData is stored as a table
  \item metadata is stored as an R object
}
}
\examples{
if (FALSE) {
    library(SummarizedExperiment)

    # Create a SE object
    se <- SummarizedExperiment(
        assays = list(counts = matrix(1:100, 10, 10)),
        colData = DataFrame(sample = paste0("S", 1:10)),
        rowData = DataFrame(gene = paste0("G", 1:10))
    )

    # Store in lake
    lake$put("my_se", se)

    # Retrieve
    se2 <- lake$get("my_se")
}

}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{SEAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SEAdapter-name}{\code{SEAdapter$name()}}
\item \href{#method-SEAdapter-can_handle}{\code{SEAdapter$can_handle()}}
\item \href{#method-SEAdapter-priority}{\code{SEAdapter$priority()}}
\item \href{#method-SEAdapter-put}{\code{SEAdapter$put()}}
\item \href{#method-SEAdapter-get}{\code{SEAdapter$get()}}
\item \href{#method-SEAdapter-components}{\code{SEAdapter$components()}}
\item \href{#method-SEAdapter-exists}{\code{SEAdapter$exists()}}
\item \href{#method-SEAdapter-list_names}{\code{SEAdapter$list_names()}}
\item \href{#method-SEAdapter-clone}{\code{SEAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SEAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-SEAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SEAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-SEAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SEAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-SEAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SEAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-SEAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SEAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-SEAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SEAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-SEAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SEAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-SEAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SEAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-SEAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SEAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SEAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
