% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_seurat.R
\name{SeuratAdapter}
\alias{SeuratAdapter}
\title{Seurat Adapter}
\value{
An R6 generator for a \code{LakeAdapter} subclass that
  serializes and restores objects of this omics layer.
}
\description{
Adapter for storing and retrieving Seurat objects.
}
\details{
This adapter stores a full-fidelity serialized Seurat object and manifest.
}
\examples{
adapter <- SeuratAdapter$new()
class(adapter)
}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{SeuratAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeuratAdapter-name}{\code{SeuratAdapter$name()}}
\item \href{#method-SeuratAdapter-priority}{\code{SeuratAdapter$priority()}}
\item \href{#method-SeuratAdapter-put}{\code{SeuratAdapter$put()}}
\item \href{#method-SeuratAdapter-get}{\code{SeuratAdapter$get()}}
\item \href{#method-SeuratAdapter-components}{\code{SeuratAdapter$components()}}
\item \href{#method-SeuratAdapter-exists}{\code{SeuratAdapter$exists()}}
\item \href{#method-SeuratAdapter-list_names}{\code{SeuratAdapter$list_names()}}
\item \href{#method-SeuratAdapter-can_handle}{\code{SeuratAdapter$can_handle()}}
\item \href{#method-SeuratAdapter-clone}{\code{SeuratAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeuratAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-SeuratAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeuratAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeuratAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-SeuratAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeuratAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeuratAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-SeuratAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeuratAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeuratAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-SeuratAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeuratAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeuratAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-SeuratAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeuratAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeuratAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-SeuratAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeuratAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeuratAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-SeuratAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeuratAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeuratAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-SeuratAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeuratAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeuratAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeuratAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeuratAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
