% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{snap}
\alias{snap}
\title{Create a snapshot of the default lake}
\usage{
snap(label, ...)
}
\arguments{
\item{label}{Label for the snapshot}

\item{...}{Additional arguments passed to Lake$snap()}
}
\value{
Invisible Lake object
}
\description{
Create a snapshot of the default lake
}
\examples{
use_lake("ex_snap", root = tempfile())
put("t", data.frame(x = 1:3))
snap("v1.0")
}
