% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_spatialexperiment.R
\name{SpatialExperimentAdapter}
\alias{SpatialExperimentAdapter}
\title{SpatialExperiment Adapter}
\description{
Adapter for storing and retrieving SpatialExperiment objects.
}
\details{
This adapter stores a full-fidelity serialized SpatialExperiment object and
manifest.
}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{SpatialExperimentAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SpatialExperimentAdapter-name}{\code{SpatialExperimentAdapter$name()}}
\item \href{#method-SpatialExperimentAdapter-priority}{\code{SpatialExperimentAdapter$priority()}}
\item \href{#method-SpatialExperimentAdapter-put}{\code{SpatialExperimentAdapter$put()}}
\item \href{#method-SpatialExperimentAdapter-get}{\code{SpatialExperimentAdapter$get()}}
\item \href{#method-SpatialExperimentAdapter-components}{\code{SpatialExperimentAdapter$components()}}
\item \href{#method-SpatialExperimentAdapter-exists}{\code{SpatialExperimentAdapter$exists()}}
\item \href{#method-SpatialExperimentAdapter-list_names}{\code{SpatialExperimentAdapter$list_names()}}
\item \href{#method-SpatialExperimentAdapter-can_handle}{\code{SpatialExperimentAdapter$can_handle()}}
\item \href{#method-SpatialExperimentAdapter-clone}{\code{SpatialExperimentAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpatialExperimentAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-SpatialExperimentAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpatialExperimentAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpatialExperimentAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-SpatialExperimentAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpatialExperimentAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpatialExperimentAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-SpatialExperimentAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpatialExperimentAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpatialExperimentAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-SpatialExperimentAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpatialExperimentAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpatialExperimentAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-SpatialExperimentAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpatialExperimentAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpatialExperimentAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-SpatialExperimentAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpatialExperimentAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpatialExperimentAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-SpatialExperimentAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpatialExperimentAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpatialExperimentAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-SpatialExperimentAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpatialExperimentAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpatialExperimentAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SpatialExperimentAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpatialExperimentAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
