% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_spectra.R
\name{SpectraAdapter}
\alias{SpectraAdapter}
\title{Spectra Adapter}
\description{
Adapter for storing and retrieving
\code{Spectra} objects.
}
\details{
This adapter preserves \code{Spectra} components:
\itemize{
  \item peak lists (\code{peaksData})
  \item per-spectrum metadata (\code{spectraData})
}
}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{SpectraAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SpectraAdapter-name}{\code{SpectraAdapter$name()}}
\item \href{#method-SpectraAdapter-can_handle}{\code{SpectraAdapter$can_handle()}}
\item \href{#method-SpectraAdapter-priority}{\code{SpectraAdapter$priority()}}
\item \href{#method-SpectraAdapter-put}{\code{SpectraAdapter$put()}}
\item \href{#method-SpectraAdapter-get}{\code{SpectraAdapter$get()}}
\item \href{#method-SpectraAdapter-components}{\code{SpectraAdapter$components()}}
\item \href{#method-SpectraAdapter-exists}{\code{SpectraAdapter$exists()}}
\item \href{#method-SpectraAdapter-list_names}{\code{SpectraAdapter$list_names()}}
\item \href{#method-SpectraAdapter-clone}{\code{SpectraAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpectraAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-SpectraAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpectraAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpectraAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-SpectraAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpectraAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpectraAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-SpectraAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpectraAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpectraAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-SpectraAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpectraAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpectraAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-SpectraAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpectraAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpectraAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-SpectraAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpectraAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpectraAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-SpectraAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpectraAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpectraAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-SpectraAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpectraAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpectraAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SpectraAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpectraAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
