% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{sql}
\alias{sql}
\title{Execute SQL on the default lake}
\usage{
sql(query, ...)
}
\arguments{
\item{query}{SQL query string}

\item{...}{Additional arguments passed to Lake$sql()}
}
\value{
Query results
}
\description{
Execute SQL on the default lake
}
\examples{
use_lake("ex_sql", root = tempfile())
put("t", data.frame(x = 1:3))
sql("SELECT COUNT(*) AS n FROM t")
}
