% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{starts_with_str}
\alias{starts_with_str}
\title{Check if string starts with a prefix}
\usage{
starts_with_str(x, prefix)
}
\arguments{
\item{x}{Character vector}

\item{prefix}{Prefix to check for}
}
\value{
Logical vector
}
\description{
Check if string starts with a prefix
}
\examples{
genes <- c("MT-CO1", "MT-CO2", "ACTB")
genes[starts_with_str(genes, "MT-")]
}
