% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{tables}
\alias{tables}
\title{List tables in the default lake}
\usage{
tables()
}
\value{
Data frame of table names
}
\description{
List tables in the default lake
}
\examples{
use_lake("ex_tables", root = tempfile())
put("t", data.frame(x = 1:3))
tables()
}
