% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_transcriptomics.R
\name{TranscriptomicsAdapter}
\alias{TranscriptomicsAdapter}
\title{Transcriptomics Adapter}
\value{
An R6 generator for a \code{LakeAdapter} subclass that
  serializes and restores objects of this omics layer.
}
\description{
Adapter for storing and retrieving transcriptomics-layer
objects.
}
\details{
This umbrella adapter is lower priority than specific adapters
(e.g. SingleCellExperiment / SummarizedExperiment) and captures
transcriptomics-marked objects.
}
\examples{
adapter <- TranscriptomicsAdapter$new()
class(adapter)
}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{TranscriptomicsAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TranscriptomicsAdapter-name}{\code{TranscriptomicsAdapter$name()}}
\item \href{#method-TranscriptomicsAdapter-priority}{\code{TranscriptomicsAdapter$priority()}}
\item \href{#method-TranscriptomicsAdapter-put}{\code{TranscriptomicsAdapter$put()}}
\item \href{#method-TranscriptomicsAdapter-get}{\code{TranscriptomicsAdapter$get()}}
\item \href{#method-TranscriptomicsAdapter-components}{\code{TranscriptomicsAdapter$components()}}
\item \href{#method-TranscriptomicsAdapter-exists}{\code{TranscriptomicsAdapter$exists()}}
\item \href{#method-TranscriptomicsAdapter-list_names}{\code{TranscriptomicsAdapter$list_names()}}
\item \href{#method-TranscriptomicsAdapter-can_handle}{\code{TranscriptomicsAdapter$can_handle()}}
\item \href{#method-TranscriptomicsAdapter-clone}{\code{TranscriptomicsAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TranscriptomicsAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-TranscriptomicsAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TranscriptomicsAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TranscriptomicsAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-TranscriptomicsAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TranscriptomicsAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TranscriptomicsAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-TranscriptomicsAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TranscriptomicsAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TranscriptomicsAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-TranscriptomicsAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TranscriptomicsAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TranscriptomicsAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-TranscriptomicsAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TranscriptomicsAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TranscriptomicsAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-TranscriptomicsAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TranscriptomicsAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TranscriptomicsAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-TranscriptomicsAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TranscriptomicsAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TranscriptomicsAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-TranscriptomicsAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TranscriptomicsAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TranscriptomicsAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TranscriptomicsAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TranscriptomicsAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
