% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{tree}
\alias{tree}
\title{Show lineage tree from the default lake}
\usage{
tree(name = NULL, ...)
}
\arguments{
\item{name}{Starting node (optional)}

\item{...}{Additional arguments passed to Lake$tree()}
}
\value{
Lineage data frame
}
\description{
Show lineage tree from the default lake
}
