% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{use_lake}
\alias{use_lake}
\title{Set or get the default lake}
\usage{
use_lake(project = NULL, ...)
}
\arguments{
\item{project}{Project name. If provided, creates/opens a lake and sets it
as default.}

\item{...}{Additional arguments passed to Lake$new()}
}
\value{
The default Lake object (invisibly when setting)
}
\description{
Set or get the default lake
}
\examples{
if (FALSE) {
    use_lake("my_analysis")
    lake <- use_lake() # Get current lake
}
}
