% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_vcf.R
\name{VCFAdapter}
\alias{VCFAdapter}
\title{VCF Adapter}
\description{
Adapter for storing and retrieving VariantAnnotation VCF objects.
}
\details{
This adapter stores a full-fidelity serialized VCF object and manifest.
}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{VCFAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VCFAdapter-name}{\code{VCFAdapter$name()}}
\item \href{#method-VCFAdapter-priority}{\code{VCFAdapter$priority()}}
\item \href{#method-VCFAdapter-put}{\code{VCFAdapter$put()}}
\item \href{#method-VCFAdapter-get}{\code{VCFAdapter$get()}}
\item \href{#method-VCFAdapter-components}{\code{VCFAdapter$components()}}
\item \href{#method-VCFAdapter-exists}{\code{VCFAdapter$exists()}}
\item \href{#method-VCFAdapter-list_names}{\code{VCFAdapter$list_names()}}
\item \href{#method-VCFAdapter-can_handle}{\code{VCFAdapter$can_handle()}}
\item \href{#method-VCFAdapter-clone}{\code{VCFAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VCFAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-VCFAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VCFAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VCFAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-VCFAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VCFAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VCFAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-VCFAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VCFAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VCFAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-VCFAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VCFAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VCFAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-VCFAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VCFAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VCFAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-VCFAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VCFAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VCFAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-VCFAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VCFAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VCFAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-VCFAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VCFAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VCFAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VCFAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VCFAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
