% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_xcms.R
\name{XCMSAdapter}
\alias{XCMSAdapter}
\title{XCMS Adapter}
\description{
Adapter for storing and retrieving xcms objects
(\code{XCMSnExp} and \code{XcmsExperiment}).
}
\details{
This adapter currently guarantees full-fidelity roundtrip by storing the
complete object and manifest as internal components.
}
\section{Super class}{
\code{\link[OmicsLake:LakeAdapter]{OmicsLake::LakeAdapter}} -> \code{XCMSAdapter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-XCMSAdapter-name}{\code{XCMSAdapter$name()}}
\item \href{#method-XCMSAdapter-can_handle}{\code{XCMSAdapter$can_handle()}}
\item \href{#method-XCMSAdapter-priority}{\code{XCMSAdapter$priority()}}
\item \href{#method-XCMSAdapter-put}{\code{XCMSAdapter$put()}}
\item \href{#method-XCMSAdapter-get}{\code{XCMSAdapter$get()}}
\item \href{#method-XCMSAdapter-components}{\code{XCMSAdapter$components()}}
\item \href{#method-XCMSAdapter-exists}{\code{XCMSAdapter$exists()}}
\item \href{#method-XCMSAdapter-list_names}{\code{XCMSAdapter$list_names()}}
\item \href{#method-XCMSAdapter-clone}{\code{XCMSAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XCMSAdapter-name"></a>}}
\if{latex}{\out{\hypertarget{method-XCMSAdapter-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XCMSAdapter$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XCMSAdapter-can_handle"></a>}}
\if{latex}{\out{\hypertarget{method-XCMSAdapter-can_handle}{}}}
\subsection{Method \code{can_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XCMSAdapter$can_handle(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XCMSAdapter-priority"></a>}}
\if{latex}{\out{\hypertarget{method-XCMSAdapter-priority}{}}}
\subsection{Method \code{priority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XCMSAdapter$priority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XCMSAdapter-put"></a>}}
\if{latex}{\out{\hypertarget{method-XCMSAdapter-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XCMSAdapter$put(lake, name, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XCMSAdapter-get"></a>}}
\if{latex}{\out{\hypertarget{method-XCMSAdapter-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XCMSAdapter$get(lake, name, ref = "@latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XCMSAdapter-components"></a>}}
\if{latex}{\out{\hypertarget{method-XCMSAdapter-components}{}}}
\subsection{Method \code{components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XCMSAdapter$components(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XCMSAdapter-exists"></a>}}
\if{latex}{\out{\hypertarget{method-XCMSAdapter-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XCMSAdapter$exists(lake, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XCMSAdapter-list_names"></a>}}
\if{latex}{\out{\hypertarget{method-XCMSAdapter-list_names}{}}}
\subsection{Method \code{list_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XCMSAdapter$list_names(lake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XCMSAdapter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-XCMSAdapter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XCMSAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
