% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{lake}
\alias{lake}
\title{Get the current default lake}
\usage{
lake()
}
\value{
The default Lake object
}
\description{
Get the current default lake
}
\examples{
if (FALSE) {
    use_lake("my_project")
    l <- lake()
    l$put("data", df)
}
}
