% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{drop}
\alias{drop}
\title{Drop data from the default lake}
\usage{
drop(name, ...)
}
\arguments{
\item{name}{Data name}

\item{...}{Additional arguments passed to Lake$drop()}
}
\value{
Invisible Lake object
}
\description{
Drop data from the default lake
}
