% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{export_data}
\alias{export_data}
\title{Export data from the default lake}
\usage{
export_data(name, path, ...)
}
\arguments{
\item{name}{Data name to export}

\item{path}{Output file path}

\item{...}{Additional arguments}
}
\value{
Invisible path
}
\description{
Export data from the default lake
}
