% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{fetch}
\alias{fetch}
\title{Read data from the default lake}
\usage{
fetch(name, ...)
}
\arguments{
\item{name}{Name of the data to read}

\item{...}{Additional arguments passed to Lake$get()}
}
\value{
The requested data
}
\description{
Note: This function is named 'fetch' to avoid conflict with base::get()
}
\seealso{
\code{\link{Lake}}
}
