% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{from}
\alias{from}
\title{Start a query from a table on the default lake}
\usage{
from(table)
}
\arguments{
\item{table}{Table name}
}
\value{
A QueryBuilder instance
}
\description{
Start a query from a table on the default lake
}
