% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{import_data}
\alias{import_data}
\title{Import data into the default lake}
\usage{
import_data(path, name, ...)
}
\arguments{
\item{path}{File path to import}

\item{name}{Name to store as}

\item{...}{Additional arguments}
}
\value{
Invisible Lake object
}
\description{
Import data into the default lake
}
