% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ol_iceberg_adapter.R
\name{ol_checkout}
\alias{ol_checkout}
\title{Restore entire project to a labeled state}
\usage{
ol_checkout(label, project = getOption("ol.project"))
}
\arguments{
\item{label}{The label name to restore to}

\item{project}{The project name}
}
\value{
Invisible TRUE on success, FALSE if no tables found
}
\description{
Restore entire project to a labeled state
}
