% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ol_commit}
\alias{ol_commit}
\title{Commit is a no-op under Iceberg (snapshots are implicit)}
\usage{
ol_commit(note = "", params = list(), project = getOption("ol.project"))
}
\description{
Commit is a no-op under Iceberg (snapshots are implicit)
}
