% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ol_drop}
\alias{ol_drop}
\title{Drop (delete) a table from the project}
\usage{
ol_drop(name, project = getOption("ol.project"))
}
\arguments{
\item{name}{Name of the table to drop}

\item{project}{Project name}
}
\description{
Drop (delete) a table from the project
}
