% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ol_drop_object}
\alias{ol_drop_object}
\title{Drop (delete) an object from the project}
\usage{
ol_drop_object(name, project = getOption("ol.project"))
}
\arguments{
\item{name}{Name of the object to drop}

\item{project}{Project name}
}
\description{
Drop (delete) an object from the project
}
