% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc.R
\name{ol_fread}
\alias{ol_fread}
\title{fread-like reader for Parquet tables}
\usage{
ol_fread(
  name,
  ref = "@latest",
  select = NULL,
  drop = NULL,
  nrows = Inf,
  filter = NULL,
  project = getOption("ol.project"),
  as_tibble = FALSE
)
}
\description{
fread-like reader for Parquet tables
}
