% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{ol_label}
\alias{ol_label}
\title{Label current state with a human-friendly alias}
\usage{
ol_label(
  label,
  state_id = NULL,
  project = getOption("ol.project"),
  .in_transaction = FALSE
)
}
\arguments{
\item{.in_transaction}{Internal parameter - if TRUE, skip transaction in child calls (caller handles it)}
}
\description{
Label current state with a human-friendly alias
}
