% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ol_list_objects}
\alias{ol_list_objects}
\title{List all saved objects in the project}
\usage{
ol_list_objects(project = getOption("ol.project"))
}
\arguments{
\item{project}{Project name}
}
\description{
List all saved objects in the project
}
