% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ol_list_tables}
\alias{ol_list_tables}
\title{List all tables in the project}
\usage{
ol_list_tables(project = getOption("ol.project"))
}
\arguments{
\item{project}{Project name}
}
\description{
List all tables in the project
}
