% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ol_log}
\alias{ol_log}
\title{Return version log for a table}
\usage{
ol_log(name = NULL, project = getOption("ol.project"))
}
\arguments{
\item{name}{Optional table name. If NULL, returns commit log.}

\item{project}{Project name}
}
\value{
Data frame of version history
}
\description{
Return version log for a table
}
