% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ol_log_commits}
\alias{ol_log_commits}
\title{View commit history}
\usage{
ol_log_commits(project = getOption("ol.project"), n = 20)
}
\arguments{
\item{project}{Project name}

\item{n}{Maximum number of commits to return}
}
\description{
View commit history
}
