% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ol_read}
\alias{ol_read}
\title{Read a table by name and reference}
\usage{
ol_read(
  name,
  ref = "@latest",
  project = getOption("ol.project"),
  collect = TRUE
)
}
\description{
Read a table by name and reference
}
