% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc.R
\name{ol_read_mae}
\alias{ol_read_mae}
\title{Compose a MultiAssayExperiment}
\usage{
ol_read_mae(
  assays,
  ref = "@latest",
  project = getOption("ol.project"),
  backing = c("hdf5", "memory")
)
}
\arguments{
\item{assays}{Named list of assay specifications}

\item{ref}{Reference string}

\item{project}{Project name}

\item{backing}{Backing mode: "hdf5" or "memory"}
}
\value{
MultiAssayExperiment object
}
\description{
Compose a MultiAssayExperiment
}
