% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc.R
\name{ol_read_mae}
\alias{ol_read_mae}
\title{Compose a MultiAssayExperiment}
\usage{
ol_read_mae(
  assays,
  ref = "@latest",
  project = getOption("ol.project"),
  backing = c("hdf5", "memory")
)
}
\description{
Compose a MultiAssayExperiment
}
