% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{ol_read_object}
\alias{ol_read_object}
\title{Read a stored object}
\usage{
ol_read_object(
  name,
  ref = "@latest",
  when = NULL,
  project = getOption("ol.project")
)
}
\arguments{
\item{name}{Name of the object to read}

\item{ref}{Reference to read from (e.g., "@latest", "@tag", "v1.0"). Defaults to "@latest"}

\item{when}{Deprecated. Use ref parameter instead. If provided, "latest" or "first"}

\item{project}{Project name}
}
\description{
Read a stored object
}
