% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ol_iceberg_adapter.R
\name{ol_read_object}
\alias{ol_read_object}
\title{Read a stored object}
\usage{
ol_read_object(
  name,
  when = c("latest", "first"),
  project = getOption("ol.project")
)
}
\description{
Read a stored object
}
