% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc.R
\name{ol_read_se}
\alias{ol_read_se}
\title{Build a SummarizedExperiment from long table}
\usage{
ol_read_se(
  name,
  ref = "@latest",
  feature_col = "feature",
  sample_col = "sample",
  value_col = "value",
  project = getOption("ol.project"),
  backing = c("hdf5", "memory")
)
}
\arguments{
\item{name}{Table name}

\item{ref}{Reference string}

\item{feature_col}{Column name containing feature ids}

\item{sample_col}{Column name containing sample ids}

\item{value_col}{Column name containing measurement values}

\item{project}{Project name}

\item{backing}{Backing mode: "hdf5" or "memory"}
}
\value{
SummarizedExperiment object
}
\description{
Build a SummarizedExperiment from long table
}
