% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc.R
\name{ol_read_se}
\alias{ol_read_se}
\title{Build a SummarizedExperiment from long table}
\usage{
ol_read_se(
  name,
  ref = "@latest",
  feature_col = "feature",
  sample_col = "sample",
  value_col = "value",
  project = getOption("ol.project"),
  backing = c("hdf5", "memory")
)
}
\description{
Build a SummarizedExperiment from long table
}
