% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ol_save}
\alias{ol_save}
\title{Save an R object via the Iceberg metadata table}
\usage{
ol_save(name, object, project = getOption("ol.project"), mime = NULL)
}
\description{
Save an R object via the Iceberg metadata table
}
