% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ol_write}
\alias{ol_write}
\title{Write a table using the Iceberg backend}
\usage{
ol_write(
  name,
  data,
  project = getOption("ol.project"),
  mode = c("create", "overwrite", "append")
)
}
\description{
Write a table using the Iceberg backend
}
