% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{ref}
\alias{ref}
\title{Get a lazy reference from the default lake}
\usage{
ref(name)
}
\arguments{
\item{name}{Table name}
}
\value{
A lazy table reference for use with dplyr
}
\description{
Get a lazy reference from the default lake
}
