% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_base.R
\name{register_adapter}
\alias{register_adapter}
\title{Register a data adapter}
\usage{
register_adapter(adapter)
}
\arguments{
\item{adapter}{An LakeAdapter instance}
}
\description{
Register a data adapter
}
