% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{sql}
\alias{sql}
\title{Execute SQL on the default lake}
\usage{
sql(query, ...)
}
\arguments{
\item{query}{SQL query string}

\item{...}{Additional arguments passed to Lake$sql()}
}
\value{
Query results
}
\description{
Execute SQL on the default lake
}
