% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{tag}
\alias{tag}
\title{Tag data in the default lake}
\usage{
tag(name, tag)
}
\arguments{
\item{name}{Data name}

\item{tag}{Tag to apply}
}
\value{
Invisible Lake object
}
\description{
Tag data in the default lake
}
